requirements.txt
# === 核心 Agent & 向量 ===
llama-index==0.10.37
llama-index-llms-openai==0.1.20
llama-index-embeddings-openai==0.1.11
llama-index-vector-stores-chroma==0.1.9

# === Ollama / 本地模型支持 ===
ollama==0.1.9

# === 向量库 ===
chromadb==0.4.24

# === HTTP & 网络（关键锁版本）===
httpx

# === 通用工具 ===
pydantic<2.0
python-dotenv
pyyaml
tqdm
